requirements.txt
Django==4.2.7
djangorestframework==3.14.0
pandas==2.1.1
matplotlib==3.8.0
seaborn==0.13.0
beautifulsoup4==4.12.2
selenium==4.15.0
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
django-filter==23.5
django-crispy-forms==2.1
crispy-bootstrap5==0.7